Separate business logic to custom hook in EmailProvider and ProductsProvider
